+++ tests/testthat/test-lines.R
@@ -6,11 +6,10 @@ context("lines")
 library(spbabel)
 sph <- sp(holey)
 spl <- as(sph, "SpatialLinesDataFrame")
-
-# test_that("lines works", {
-#   expect_that(anglr(spl), is_a("linemesh"))
-#   expect_that(anglr(sp::geometry(spl)), is_a("linemesh"))
-#   
-# })
+ test_that("lines works", {
+   expect_that(SC(spl), is_a("SC"))
+   expect_that(SC(sp::geometry(spl)), is_a("SC"))
+   
+})