+++ tests/testthat/test-sanity.R
@@ -20,6 +20,12 @@ test_that("sanity prevails 3", {
   expect_silent(plot3d(TRI(minimal_mesh)))
 })
 test_that("sanity prevails 4", {
+  z <- copy_down(DEL(minimal_mesh), gebco1)
+  expect_equivalent(z$vertex$z_, c(-4890, -4916.05, -4828.88, -4795.832, -4846.382, -4813.1744, 
+                                   -4841.022, -4877.1168, -4849.578, -4846.94, -4862.78, -4875.4552, 
+                                   -4772.29496, -4758.55736))
+  expect_message(bb <- copy_down(SC(cont_tas), gebco1), "transforming")
+  expect_true(all(bb$vertex$z_ > 150 & bb$vertex$z_ < 200))
   expect_silent(copy_down(DEL(minimal_mesh), gebco1))
   expect_silent(copy_down(TRI(minimal_mesh), gebco1))
   expect_silent(copy_down(DEL(minimal_mesh), "a"))